@@ -11,55 +11,74 @@ import { Router } from '@angular/router';
 })
 export class ProductListComponent implements OnInit {
 
-  products: Product[]=[];
-  currentCategoryId: number=1;
-  searchMode: boolean=false;
+  products: Product[] = [];
+  currentCategoryId: number = 1;
+  previousCategoryId: number = 1;
+  searchMode: boolean = false;
+
+  thePageNumber: number = 1;
+  thePageSize: number = 1;
+  theTotalElements: number = 0;
 
   constructor(private productService: ProductService,
-  private route: ActivatedRoute,private router: Router ) { }
+    private route: ActivatedRoute, private router: Router) { }
 
   ngOnInit(): void {
-  this.route.paramMap.subscribe(()=>{
-  this.listProducts();
-  });
+    this.route.paramMap.subscribe(() => {
+      this.listProducts();
+    });
   }
 
-  listProducts(){
-  this.searchMode = this.route.snapshot.paramMap.has('keyword');
-  if(this.searchMode){
-  this.handleSearchProducts();
-  }else{
-  this.handleListProducts();
-  }
-  }
-
-  handleSearchProducts(){
-   const theKeyword: string = this.route.snapshot.paramMap.get('keyword');
-
-   this.productService.getSearchList(theKeyword).subscribe(
-   data=> {
-   this.products = data;
-   }
-   )
-  }
-
-  handleListProducts(){
-  const categoryId: boolean = this.route.snapshot.paramMap.has('id');
-
-    if(categoryId){
-     this.currentCategoryId = +this.route.snapshot.paramMap.get('id')!;
-    }else{
-    this.currentCategoryId = 1;}
-
-    this.productService.getProductList(this.currentCategoryId).subscribe(
-    data=> {
-    this.products = data;
+  listProducts() {
+    this.searchMode = this.route.snapshot.paramMap.has('keyword');
+    if (this.searchMode) {
+      this.handleSearchProducts();
+    } else {
+      this.handleListProducts();
     }
+  }
+
+  handleSearchProducts() {
+    const theKeyword: string = this.route.snapshot.paramMap.get('keyword');
+
+    this.productService.getSearchList(theKeyword).subscribe(
+      data => {
+        this.products = data;
+      }
     )
   }
 
-  doSearch(val: string){
-  this.router.navigateByUrl(`product/getInfo/${val}`);
+  handleListProducts() {
+    const categoryId: boolean = this.route.snapshot.paramMap.has('id');
+
+    if (categoryId) {
+      this.currentCategoryId = +this.route.snapshot.paramMap.get('id')!;
+    } else {
+      this.currentCategoryId = 1;
+    }
+
+
+  if(this.previousCategoryId != this.currentCategoryId) {
+  this.thePageNumber = 1;
+   }
+   this.previousCategoryId = this.currentCategoryId;
+
+    this.productService.getProductListPaginate(this.thePageNumber - 1,
+      this.thePageSize,
+      this.currentCategoryId)
+      .subscribe(
+        data => {
+          this.products = data.content;
+          this.thePageNumber = data.number + 1;
+          this.thePageSize = data.size;
+          this.theTotalElements = data.totalPages;
+        }
+      )
   }
+
+
+ doSearch(value: string){
+  this.router.navigateByUrl(`product/getInfo/${value}`);
+}
 
 }
